プロジェクトを削除するとき、日報が存在した場合削除できない仕様に変更

+++ app/Http/Controllers/Projects/ProjectController.php
@@ -71,7 +71,11 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        return response()->delete($project->delete());
+        if((int)$project->dailies()->count() === 0) {
+            return response()->delete($project->delete());
+        } else {
+            return redirect()->back()->withErrors('このプロジェクトには日報が存在するため、削除できません');
+        }
     }
 
     /**
